Split models in seperated modules
