+++ src/kikinzage/client/asyncio.py
@@ -9,8 +9,8 @@ from httpx import USE_CLIENT_DEFAULT
 from httpx._client import UseClientDefault
 
 from .. import models
-from ..models import Formaat
 from ..models import ResponseType
+from ..models.enum import Formaat
 from .base import KikinzageBaseClient
